refactor(servidor): crear un trait para la estructura del producto
Permite compartir la estructura entre diferentes tests
